requirements.txt
numpy
opencv_python
opencv_python_headless
Roboflow
Pillow
PyAutoGUI
pynput
pytesseract
torch
python-Levenshtein
PyQt5
ultralytics
